pyproject.toml
[tool.poetry]
name = "playbot"
version = "0.1.0"
description = "RobotFramework wrapper of Playwright-Python"
authors = ["radekBednarik <radek.bednarik@tesena.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
robotframework = "^4.0.3"
playwright = "^1.12.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pylint = "^2.8.3"
mypy = "^0.902"
isort = "^5.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
